changes shift_manager to module
